+++ content_scripts/autohalt-youtube.js
@@ -4,21 +4,25 @@ function findAutoplayButton() {
 }
 
 function disableAutoplay() {
+    const result = {found: false, clicked: false};
+
     const autoplayButton = findAutoplayButton();
     if (!autoplayButton) {
-        // console.debug("Autoplay button not found");
-        return false;
+        return result;
     }
-    // console.debug("Found autoplay button:", autoplayButton);
+    result.found = true;
 
     if (autoplayButton.getAttribute('aria-pressed') === 'true') {
         console.info("Clicking autoplay button");
         autoplayButton.click();
+        result.clicked = true;
     }
-    return true;
+
+    return result;
 }
 
 // YouTube remembers the user's autoplay selection as they navigate around, so
-// we can stop listening for mutation events as soon as the up next toggle
-// appears.
-setupAutoHalt('youtube', disableAutoplay, {disconnectOnSuccess: true});
+// we can stop listening for mutation events as soon as we find up next toggle
+// (and turn it off if needed). This allows the user to manually turn it back
+// on if desired.
+setupAutoHalt('youtube', disableAutoplay, {disconnectOnFound: true});